morse: Morse code player module.
This is a code module that can play arbitrary morse-code symbols using
the 'sfx' module to control tones.  The module can be set to any
frequency and speed (the time of a "dit" in milliseconds).

At present, I support the letters and digits, no punctuation, but those
can be added, suitable tables will need to be defined for those.

git-svn-id: https://svn.code.sf.net/p/freetel/code/codec2-dev@2356 01035d8c-6547-0410-b346-abe4f91aad63